Move gallery to tools so we can use it in an installed kivy (during tests).
