+++ kivy/tools/gallery.py
